backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pdfplumber>=0.10.0
python-dotenv>=1.0.0
groq>=0.4.0
deep-translator>=1.11.0
Pillow>=10.0.0
httpx<0.28.0
